+++ CourselyLabs-front/src/router/index.ts
@@ -1,8 +1,9 @@
 //import { route } from 'quasar/wrappers';
-import { 
+import {
   createRouter,
   createWebHistory, } from 'vue-router';
 import routes from './routes';
+import { useAuthStore } from '../stores/auth';
 //import HomeView from '../views/HomeView.vue'
 
 /* const router = createRouter({
@@ -40,6 +41,15 @@ import routes from './routes';
 const router = createRouter({
   history: createWebHistory(import.meta.env.BASE_URL),
   routes,
+});
+
+router.beforeEach((to) => {
+  if (to.meta.requiresAuth) {
+    const authStore = useAuthStore();
+    if (!authStore.isLoggedIn) {
+      return { path: '/login' };
+    }
+  }
 });
 
 export default router;