+++ CourselyLabs-front/src/router/routes.ts
@@ -13,6 +13,11 @@ const routes: RouteRecordRaw[] = [
         { path: 'login' , component: () => import('../views/formView.vue') },
         { path: 'iniSession', component: () => import('../components/formLogin.vue') },
         { path: 'register' , component: () => import('../components/formRegister.vue') },
+        {
+          path: 'profile',
+          component: () => import('../views/ProfileView.vue'),
+          meta: { requiresAuth: true },
+        },
         ],
     },